refactor: split logger into base + client-specific derived classes
- include/common/logger.h: abstract base — pure virtual print(),
  common resultScope* / emit / recordSkip, protected onMetricEmitted() hook
- include/cli/logger_cli.h + src/cli/logger_cli.cpp: LoggerCli (stdout, baseline)
- android/.../logger_android.h + logger_android.cpp: LoggerAndroid (JNI callbacks)
- No platform-specific #ifdef in common headers
- Peak no longer creates logger — caller sets it before applyOptions()
- CLI main creates LoggerCli, Android entry creates LoggerAndroid
- src/common/logger.cpp: base class implementation
- Future clients (iOS etc.) just add a new derived class
