+++ src/common/peak.cpp
@@ -1,10 +1,7 @@
 #include <common/peak.h>
 #include <common/options.h>
 
-Peak::Peak()
-    : log(new logger())
-{
-}
+Peak::Peak() = default;
 
 void Peak::applyOptions(const CliOptions &opts)
 {
@@ -13,9 +10,4 @@ void Peak::applyOptions(const CliOptions &opts)
     warmupCount    = opts.warmupCount;
     targetTimeUs   = opts.targetTimeUs;
     gating.copyFrom(opts);
-
-    // Per-backend loggers handle stdout + baseline-compare deltas only.
-    // File output is centralized in src/cli/main.cpp after all backends
-    // have run, so a single dump file aggregates every backend's rows.
-    log.reset(new logger(opts.compareFile));
 }
